public

+++ DistSystProject2/DistSystProject2/ZAB.cs
@@ -267,9 +267,9 @@ namespace Server
                 using (TcpClient client = new TcpClient(tcp.dns, tcp.port))
                 {
                     TCP t = new TCP(client);
-                    var text = String.Format("proposal {0} {1} {2}", p.z.epoch, p.z.counter, p.v);
+                    var text = String.Format("proposal {0} {1} {2}", (int)p.z.epoch, (int)p.z.counter, p.v);
                     t.sendTcpMessage(text);
-                    Console.WriteLine("Sent proposal ({0}, {1}, '{2}') to {3}", p.z.epoch, p.z.counter, p.v, tcp.dns);
+                    Console.WriteLine("Sent proposal ({0}, {1}, '{2}') to {3}", (int)p.z.epoch, (int)p.z.counter, p.v, tcp.dns);
                 }
             }
             catch (Exception ex)
